refactored the profile page as well as the authRoute to be able to check the session in the server components
