+++ app/profile/[id]/page.jsx
@@ -1,42 +1,26 @@
-'use client';
-import React, { useState, useEffect } from 'react';
 import Image from 'next/image';
-import { useSession } from 'next-auth/react';
 import Link from 'next/link';
+import dbConnect from '@/utils/dbConnect';
+import Resource from '@/models/Resource';
+import { getServerSession } from 'next-auth/next';
+import { authOptions } from '@/app/lib/auth';
 
-const Profile = () => {
-  const { data: session } = useSession();
-  const [githubUserDetails, setGithubUserDetails] = useState();
-  const [unpublishedResources, setUnpublishedResources] = useState();
+const Profile = async () => {
+  const session = await getServerSession(authOptions);
+  await dbConnect();
+  const foundResources = await Resource.find();
+  const filteredResources = foundResources.filter((resource) => {
+    return resource.published !== true;
+  });
 
-  async function getData() {
-    const res = await fetch('/api/resource');
-    const data = await res.json();
-    console.log('data------___________', data);
-    const filteredResources = data.filter((resource) => {
-      return resource.published !== true;
-    });
-    setUnpublishedResources(filteredResources);
-  }
-
-  useEffect(() => {
-    getData();
-  }, []);
-
-  // console.log('session-------++------', session);
-
-  useEffect(() => {
-    const getGithubDetails = async () => {
-      const response = await fetch(
-        `https://api.github.com/user/${session?.user?.githubId}}`
-      );
-      const UserGitubDetails = await response.json();
-      setGithubUserDetails(UserGitubDetails);
-    };
-    getGithubDetails();
-  }, [session]);
-
-  console.log('unpublishedResources++++++++++++', unpublishedResources);
+  const getGithubDetails = async () => {
+    const response = await fetch(
+      `https://api.github.com/user/${session?.user?.githubId}}`
+    );
+    const UserGitubDetails = await response.json();
+    return UserGitubDetails;
+  };
+  const githubUserDetails = await getGithubDetails();
 
   if (session) {
     return (
@@ -50,7 +34,7 @@ const Profile = () => {
         {session.user.admin && (
           <>
             <h2>unpublished resources</h2>
-            {unpublishedResources?.map((resource) => {
+            {filteredResources?.map((resource) => {
               return (
                 <Link key={resource._id} href={`/create/${resource._id}`}>
                   <h2>{resource.title}</h2>
